include: bluetooth: move ble_date_time.h
Move ble_date_time.h from bm/bluetooth/services to bm/bluetooth.

Currently it is only a helper file for the CGM service and not
strictly a service itself. The ble_date_time.h header now resides
in the same location as the ble_racp.h, which is of similar use.
