+++ subsys/bluetooth/services/ble_cgms/cgms_sst.h
@@ -21,7 +21,7 @@
 #define BLE_CGMS_SST_H__
 
 #include <ble.h>
-#include <bm/bluetooth/services/ble_date_time.h>
+#include <bm/bluetooth/ble_date_time.h>
 #include <bm/bluetooth/services/ble_cgms.h>
 
 #ifdef __cplusplus
